requirements.txt
numpy==2.1.1
opencv-python==4.11.0.86
pillow==11.1.0
scipy==1.15.2
torch==2.6.0
torchvision==0.21.0
ultralytics==8.3.94
